Application/requirements.txt
absl-py==0.12.0
altgraph==0.17
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
auto-py-to-exe==2.7.11
backcall==0.2.0
batchgenerators==0.23
beautifulsoup4==4.10.0
bleach==4.1.0
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docutils==0.16
Eel==0.12.4
efficientnet==1.1.1
efficientnet-pytorch==0.7.1
einops==0.4.1
entrypoints==0.4
executing==0.8.3
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gevent==21.1.2
gevent-websocket==0.10.1
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imageio==2.16.1
import-ipynb==0.1.3
ipykernel==6.9.2
ipynb==0.5.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.0.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
Keras==2.4.3
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kiwisolver==1.3.1
linecache2==1.0.0
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
MedPy==0.4.0
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.10
numpy==1.22.3
oauthlib==3.1.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
panda3d==1.10.8
pandas==1.2.1
pandocfilters==1.5.0
parso==0.8.3
pefile==2019.4.18
perlin-noise==1.8
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.15.8
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==2.0.1
Pygments==2.7.4
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
pyrsistent==0.18.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.3.0
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.19.2
scikit-learn==0.24.1
scipy==1.6.0
screeninfo==0.6.7
Send2Trash==1.8.0
SimpleITK==2.1.1
six==1.15.0
sklearn==0.0
soupsieve==2.3.1
stack-data==0.2.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorboardX==2.5
tensorflow==2.5.0rc1
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==2.1.0
tifffile==2022.3.25
timm==0.5.4
torch==1.8.1+cu102
torchaudio==0.8.1
torchvision==0.9.1+cu102
tornado==6.1
tqdm==4.63.1
traceback2==1.4.0
traitlets==5.1.1
typing-extensions==3.7.4.3
unittest2==1.1.0
urllib3==1.26.3
ursina==3.3.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==3.6.0
wrapt==1.12.1
yacs==0.1.8
zope.event==4.5.0
zope.interface==5.2.0
